Refactor plot_graph to create_gif
